pyproject.toml
[tool.poetry]
name = "serial_bus"
version = "0.1.0"
description = "A Python package that standardizes data parsing from multiple serialized formats into pydantic models."
authors = ["André Lima"]
documentation= "https://github.com/theandrelima/serial-weaver"
keywords=["serialization", "deserialization", "parsing", "pydantic-models", "seria-to-pydantic", "pydantic-to-serial"]
license = "MIT"
readme = "README.md"
packages = [{include = "serial_bus"}]


[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.6.1"
pyyaml = "^6.0.1"
sortedcontainers = "^2.4.0"
jinja2 = "^3.1.3"
toml = "^0.10.2"
email-validator = "^2.1.0.post1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
mdformat = "^0.7.17"
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
